requirements.txt
requests>=2.31
python-dotenv>=1.0
numpy>=1.24
scikit-learn>=1.3
tensorflow>=2.15
librosa>=0.10
soundfile>=0.12
tqdm>=4.65
